Move feedback buttons out to their own component
